requirements.txt
django-celery==3.1.17
django-cleanup==8.1.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-utils-six==2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pillow==10.2.0
pip-chill==1.0.3
psycopg2==2.9.9
python-dotenv==1.0.1
